requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.4
django-axes==8.0.0
django-cors-headers==4.7.0
django-filter==25.1
django-phonenumber-field==8.1.0
django-ratelimit==4.0.0
django-redis==6.0.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djoser==2.2.2
drf-nested-routers==0.94.2
drf-spectacular==0.28.0
fakeredis==2.30.1
frozenlist==1.7.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
multidict==6.6.3
packaging==25.0
phonenumbers==9.0.9
pillow==11.3.0
prompt_toolkit==3.0.51
propcache==0.3.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
pyclamd==0.4.0
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
smsir-python==1.0.8
sortedcontainers==2.4.0
sqlparse==0.5.3
django-sslserver==0.22
twilio==9.6.5
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1
